requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0
```

**`.gitignore`:**
```
.env
__pycache__/
*.pyc
dlq.json
*.log
.idea/
.vscode/